Add files via upload
About Dataset

This is the most recent data on the internet, of the used cars along with their prices to be sold in the local Pakistani market. This data is being scraped from the Pakisatns no. 1 used car selling website Pakwheels. In this data set, we have multiple features which play an important role in suggesting and evaluating the price of the specific car. In this dataset, we have car models from 1990 till 2021 including more than 280 unique model variants. As most of the available Pakistan used cars datasets were outdated and have incomplete information with mixed up labels and attributes, we created our own dataset using web scraping technique in python to extract data of used vehicles from www.pakwheels.com. Our dataset is updated with all appropriate and significant information required for prediction and we have cleaned null value as well as less-significant information provided by the user of the vehicle. The features available in this dataset are Make, Model, Version, Price, Make Year, CC, Assembly, Mileage, Registered City and Transmission. With 72656 examples.